What is running on the nano at the moment
This is a gross specimen seems to work as intended. I was inspired by the modularity of leadership's code to have separate functions for movements. The readUltrasonicDistance function is from TinkerCAD (who got it from http://www.arduino.cc/en/Tutorial/Ping as public domain code).  Everything is in one ino rather than having a main ino with calls to separate files (because I am a luddite and don't trust computers). Still need to test out if the threshold distances (that decide whether or not to make a turn) are valid, and it also might be cool to have a decision to go backwards if too close to a front wall.